Move ReactQueryClientProvider from app
