+++ modules/query/queryClient.ts
@@ -1,0 +1,7 @@
+import { QueryClient } from '@tanstack/react-query'
+
+export const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: { retry: false },
+  },
+})
