+++ modules/user/queries.tsx
@@ -8,7 +8,6 @@
 import { UserType } from '@stanfordbdhg/engagehf-models'
 import { queryOptions } from '@tanstack/react-query'
 import { query, where } from 'firebase/firestore'
-import { queryClient } from '@/app/ReactQueryClientProvider'
 import { docRefs, getCurrentUser, refs } from '@/modules/firebase/app'
 import { type Invitation, type Organization } from '@/modules/firebase/models'
 import { mapAuthData } from '@/modules/firebase/user'
@@ -18,6 +17,7 @@ import {
   getDocsData,
   type UserAuthenticationInformation,
 } from '@/modules/firebase/utils'
+import { queryClient } from '@/modules/query/queryClient'
 
 export const getNonAdminInvitations = (organizationIds: string[]) =>
   query(
