+++ routes/~__root.tsx
@@ -9,10 +9,10 @@ import '../app/globals.css'
 import { createRootRoute, Outlet, redirect } from '@tanstack/react-router'
 import { NextIntlClientProvider } from 'next-intl'
 import { useLayoutEffect } from 'react'
-import { ReactQueryClientProvider } from '@/app/ReactQueryClientProvider'
 import { auth } from '@/modules/firebase/app'
 import { AuthProvider } from '@/modules/firebase/AuthProvider'
 import messages from '@/modules/messages/translations/en.json'
+import { ReactQueryClientProvider } from '@/modules/query/ReactQueryClientProvider'
 import { routes } from '@/modules/routes'
 import { lightTheme } from '@/packages/design-system/src'
 import { Toaster } from '@/packages/design-system/src/components/Toaster'
