+++ routes/~_dashboard/~users/~$id.tsx
@@ -9,9 +9,9 @@ import { updateDoc } from '@firebase/firestore'
 import { UserType } from '@stanfordbdhg/engagehf-models'
 import { createFileRoute, notFound, useRouter } from '@tanstack/react-router'
 import { Users } from 'lucide-react'
-import { queryClient } from '@/app/ReactQueryClientProvider'
 import { callables, docRefs, ensureType } from '@/modules/firebase/app'
 import { getDocDataOrThrow } from '@/modules/firebase/utils'
+import { queryClient } from '@/modules/query/queryClient'
 import {
   getUserData,
   userOrganizationQueryOptions,
